+++ game/code/player/zombie/status_effect/effects.lua
@@ -2,8 +2,9 @@ local Entangle = require('code.player.status_effect.entangle')
 local Burn = require('code.player.zombie.status_effect.burn')
 local Track = require('code.player.zombie.status_effect.track')
 local Hide = require('code.player.zombie.status_effect.hide')	  
+local Scanned = require('code.player.zombie.status_effect.scanned')
 
-local effects = {Entangle, Burn, Track, Hide}
+local effects = {Entangle, Burn, Track, Hide, Scanned}
 
 for _, Class in ipairs(effects) do
   effects[Class.name] = Class